new it translation
new selection mode called normal, where only the the boxes left from the current one are influenced.

fix crash when crossfading is on and the driver is changed while playing. the problem wa that in this case the sound engine is closed while a song is not stopped but faded out. now in the case of engine stop the playing song is stopped too.