fix some warnings
add some code for generic equalizer support in musikengine.

show text which tells the user that the engine does not support equalizer if the engine does not support it 

add code to remember changes to songidcache during a transaction and replay these changes. this is needed because while in a transaction the changes are not visiable to queries in the main thread.

fix a bug in MusikTagFrame where m_Dirty was set false on each new population of the dialog, so all changes before were forgotten.